Fix probing servers for scopes, and fix other mismatch errors when loading hosts. Fixes #382
Some enterprise servers may not have the /rate_limit path. This is fine for checking whether
the server is a GitHub instance (the headers will be returned regardless), but it doesn't work
for checking scopes. Since the user:read scope is mandatory, valid credentials will work with
the /user endpoint. This has the added benefit that we can also grab the username in the
process so that we can show it elsewhere (not a part of this patch)