Treat write:discussion scope as optional when validating against GHE hosts. Fixes #507
The write:discussion scope is a newer scope that older versions of GHE don't have
(or have under a different name). Most of the features of the extension don't require
this scope (it's mainly for team discussions), so it won't break functionality if
this scope is optional when validating scopes.

This only affects validating personal access tokens created manually by the user for
GHE hosts (not github.com), as the sign in process is only available in versions of
GHE that also include the write:discussion scope.