Disable scheduled pre-release out of main (#4984)
We now have more people working in this repo, so it's easier to branch for release during endgame rather ask everyone to hold off on changes while we stabilize.
To that end, I've branched for release (`release/0.68`) and with this PR the nightly pre-release build out of `main` will be disabled. This will need to be re-enabled after release.