Disable copy commands in embedded editors
This disables these commands in the interactive playground and chat blocks

In the future, it would be better to instead check that the file belongs to the github repo. That way we'd disable this command in files from virtual file systems. However that is a larger change and would require a custom context key