Update to use eslint9
Updates the package to use eslint9. This switches to a flat config file and also updates a few dependencies

I generally tried to keep the existing logic and rules with this change. We can revisit these later

I did end up having to make a few code changes for new errors that showed up after updating