Update to use eslint9 (#7876)
* Update to use eslint9

Updates the package to use eslint9. This switches to a flat config file and also updates a few dependencies

I generally tried to keep the existing logic and rules with this change. We can revisit these later

I did end up having to make a few code changes for new errors that showed up after updating

* Update all pipelines to node 20

The release pipelines should already be using this but the pr pipeline was on an old version

* Update common-setup instead

* Add back missing rules

* Fix webviews rule

* Remove some bad eslint comments

---------

Co-authored-by: Alex Ross <38270282+alexr00@users.noreply.github.com>